requirements.txt
Flask>=2.3
mysql-connector-python>=8.1
python-dotenv>=1.0
